+++ src/features/authentication/hooks/useWhenLoggedIn.ts
@@ -3,14 +3,14 @@ import { useEffect } from 'react'
 import { useAppSelector } from 'hooks/redux'
 import { isUserLoggedIn } from 'store'
 
-const useWhenLoggedIn = (callback = () => {}, properties: any = []) => {
+const useWhenLoggedIn = (callback = () => {}) => {
   const isLoggedIn = useAppSelector(isUserLoggedIn)
 
   useEffect(() => {
     if (isLoggedIn) {
-      console.log('isLoggedIn')
       callback()
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isLoggedIn])
 }
